warning fixes
git-svn-id: svn://svn.berlios.de/open-iscsi@63 d7303112-9cec-0310-bdd2-e83a94d6c2b6

+++ usr/discovery.c
@@ -453,11 +453,11 @@ process_sendtargets_response(struct string_buffer *sendtargets,
 			 * beginning of the buffer.
 			 */
 			log_debug(7,
-				 "processed %ld bytes of sendtargets data, "
-				 "%ld remaining",
-				 record - buffer_data(sendtargets),
-				 buffer_data(sendtargets) +
-				 data_length(sendtargets) - record);
+				 "processed %d bytes of sendtargets data, "
+				 "%d remaining",
+				 (int)(record - buffer_data(sendtargets)),
+				 (int)(buffer_data(sendtargets) +
+				 data_length(sendtargets) - record));
 			remove_initial(sendtargets,
 				       record - buffer_data(sendtargets));
 		}
